Refactoring
General cleanup. Added a separate strings processing file. Changed the way some uci commands are sent.
